Separate test for find_missing_number
